pyproject.toml
[tool.poetry]
name = "fulgens-cli"
version = "0.1.0"
description = ""
authors = ["Jon Friedman <jonathan.friedman@freenome.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-gitlab = "^2.10.1"
click = "^8.0.3"

[tool.poetry.dev-dependencies]
ipython = "^7.30.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
